Refactor: modularize models and records into separate files
Split multi-type files into dedicated files for models, records, enums, and exceptions across Executor, FileUploader, Fitness, Paissa, Shift, and test utilities. Added explicit namespaces and improved property initialization. No functional changes; improves code organization, clarity, and maintainability.
