+++ src/service-defaults/Tomeshelf.ServiceDefaults/ExecutorDiscoveryConstants.cs
@@ -1,0 +1,12 @@
+namespace Tomeshelf.ServiceDefaults;
+
+/// <summary>
+///     Constants for executor discovery endpoints.
+/// </summary>
+public static class ExecutorDiscoveryConstants
+{
+    /// <summary>
+    ///     Default discovery route path.
+    /// </summary>
+    public const string DefaultPath = "/.well-known/executor-endpoints";
+}
