+++ src/services/fileuploader/Tomeshelf.FileUploader.Application/Features/Uploads/Models/BookUploadResult.cs
@@ -1,0 +1,3 @@
+namespace Tomeshelf.FileUploader.Application.Features.Uploads.Models;
+
+public sealed record BookUploadResult(string BundleName, string BookTitle, int FilesUploaded, int FilesSkipped);
