+++ src/services/fileuploader/Tomeshelf.FileUploader.Infrastructure/Upload/BookFile.cs
@@ -1,0 +1,3 @@
+namespace Tomeshelf.FileUploader.Infrastructure.Upload;
+
+public sealed record BookFile(string FullPath, string TargetFileName, long Length);
