+++ src/services/fileuploader/Tomeshelf.FileUploader.Infrastructure/Upload/BookPlan.cs
@@ -1,0 +1,5 @@
+using System.Collections.Generic;
+
+namespace Tomeshelf.FileUploader.Infrastructure.Upload;
+
+public sealed record BookPlan(string BundleName, string BookTitle, IReadOnlyList<BookFile> Files);
