+++ src/shared/Tomeshelf.Application/DummyService.cs
@@ -1,0 +1,3 @@
+﻿namespace Tomeshelf.Application.Shared;
+
+public class DummyService : IDummyService { }
