+++ src/shared/Tomeshelf.Application/IDummyService.cs
@@ -1,0 +1,3 @@
+﻿namespace Tomeshelf.Application.Shared;
+
+public interface IDummyService { }
