+++ src/web/Tomeshelf.Executor/Configuration/ExecutorOptions.cs
@@ -1,7 +1,5 @@
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Net.Http;
-using Tomeshelf.Executor.Validation;
 
 namespace Tomeshelf.Executor.Configuration;
 
@@ -19,28 +17,4 @@ public class ExecutorOptions
     /// </summary>
     [Required]
     public List<EndpointScheduleOptions> Endpoints { get; set; } = [];
-}
-
-public class EndpointScheduleOptions
-{
-    [Required]
-    public string Name { get; set; } = string.Empty;
-
-    [Required]
-    [AbsoluteUrl]
-    public string Url { get; set; } = string.Empty;
-
-    [Required]
-    public string Cron { get; set; } = string.Empty;
-
-    public bool Enabled { get; set; } = true;
-
-    public string Method { get; set; } = HttpMethod.Post.Method;
-
-    /// <summary>
-    ///     Optional TimeZone identifier understood by Quartz (defaults to UTC).
-    /// </summary>
-    public string? TimeZone { get; set; }
-
-    public Dictionary<string, string>? Headers { get; set; }
 }
