+++ src/web/Tomeshelf.Executor/Services/ExecutorConfigurationStore.cs
@@ -1,7 +1,5 @@
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
@@ -13,11 +11,11 @@ namespace Tomeshelf.Executor.Services;
 
 public sealed class ExecutorConfigurationStore : IExecutorConfigurationStore
 {
-    private static readonly JsonSerializerOptions SerializerOptions = new JsonSerializerOptions(JsonSerializerDefaults.Web) { WriteIndented = true };
+    private static readonly JsonSerializerOptions SerializerOptions = new(JsonSerializerDefaults.Web) { WriteIndented = true };
 
     private readonly string _defaultFilePath;
     private readonly string? _environmentFilePath;
-    private readonly SemaphoreSlim _lock = new SemaphoreSlim(1, 1);
+    private readonly SemaphoreSlim _lock = new(1, 1);
     private readonly ILogger<ExecutorConfigurationStore> _logger;
 
     public ExecutorConfigurationStore(IHostEnvironment environment, ILogger<ExecutorConfigurationStore> logger)
@@ -113,38 +111,6 @@ public sealed class ExecutorConfigurationStore : IExecutorConfigurationStore
             Executor = options.Clone();
         }
 
-        public ExecutorOptions Executor { get; set; } = new ExecutorOptions();
-    }
-}
-
-internal static class ExecutorOptionsExtensions
-{
-    public static ExecutorOptions Clone(this ExecutorOptions source)
-    {
-        var clone = new ExecutorOptions
-        {
-            Enabled = source.Enabled,
-            Endpoints = source.Endpoints
-                              .Select(Clone)
-                              .ToList()
-        };
-
-        return clone;
-    }
-
-    private static EndpointScheduleOptions Clone(EndpointScheduleOptions source)
-    {
-        return new EndpointScheduleOptions
-        {
-            Name = source.Name,
-            Url = source.Url,
-            Method = source.Method,
-            Cron = source.Cron,
-            Enabled = source.Enabled,
-            TimeZone = source.TimeZone,
-            Headers = source.Headers is null
-                ? null
-                : new Dictionary<string, string>(source.Headers, StringComparer.OrdinalIgnoreCase)
-        };
+        public ExecutorOptions Executor { get; set; } = new();
     }
 }
