+++ src/web/Tomeshelf.Web/Models/Bundles/BundleUploadResultModel.cs
@@ -17,14 +17,3 @@ public sealed class BundleUploadResultModel
 
     public IReadOnlyList<BookUploadResultModel> Books { get; init; } = Array.Empty<BookUploadResultModel>();
 }
-
-public sealed class BookUploadResultModel
-{
-    public string BundleName { get; init; } = string.Empty;
-
-    public string BookTitle { get; init; } = string.Empty;
-
-    public int FilesUploaded { get; init; }
-
-    public int FilesSkipped { get; init; }
-}
