+++ src/web/Tomeshelf.Web/Models/Fitness/FitbitActivityModel.cs
@@ -1,0 +1,10 @@
+namespace Tomeshelf.Web.Models.Fitness;
+
+public sealed class FitbitActivityModel
+{
+    public int? Steps { get; set; }
+
+    public double? DistanceKm { get; set; }
+
+    public int? Floors { get; set; }
+}
