+++ src/web/Tomeshelf.Web/Models/Fitness/FitbitSleepLevelsModel.cs
@@ -1,0 +1,12 @@
+namespace Tomeshelf.Web.Models.Fitness;
+
+public sealed class FitbitSleepLevelsModel
+{
+    public int? DeepMinutes { get; set; }
+
+    public int? LightMinutes { get; set; }
+
+    public int? RemMinutes { get; set; }
+
+    public int? WakeMinutes { get; set; }
+}
