+++ src/web/Tomeshelf.Web/Models/Paissa/PaissaDistrictModel.cs
@@ -1,0 +1,5 @@
+using System.Collections.Generic;
+
+namespace Tomeshelf.Web.Models.Paissa;
+
+public sealed record PaissaDistrictModel(int Id, string Name, IReadOnlyList<PaissaSizeGroupModel> Tabs);
