+++ src/web/Tomeshelf.Web/Models/Paissa/PaissaPlotModel.cs
@@ -1,0 +1,5 @@
+using System;
+
+namespace Tomeshelf.Web.Models.Paissa;
+
+public sealed record PaissaPlotModel(int Ward, int Plot, long Price, int Entries, DateTimeOffset LastUpdatedUtc, bool AllowsPersonal, bool AllowsFreeCompany, bool IsEligibilityUnknown);
