+++ src/web/Tomeshelf.Web/Models/Paissa/PaissaSizeGroupModel.cs
@@ -1,0 +1,5 @@
+using System.Collections.Generic;
+
+namespace Tomeshelf.Web.Models.Paissa;
+
+public sealed record PaissaSizeGroupModel(string Size, string SizeKey, IReadOnlyList<PaissaPlotModel> Plots);
