+++ src/web/Tomeshelf.Web/Models/Shift/RedeemErrorCode.cs
@@ -1,0 +1,12 @@
+namespace Tomeshelf.Web.Models.Shift;
+
+public enum RedeemErrorCode
+{
+    InvalidCredentials,
+    CsrfMissing,
+    NoRedemptionOptions,
+    RedemptionFailed,
+    NetworkError,
+    AccountMisconfigured,
+    Unknown
+}
