+++ src/web/Tomeshelf.Web/Models/Shift/RedeemRequestModel.cs
@@ -1,0 +1,3 @@
+namespace Tomeshelf.Web.Models.Shift;
+
+public sealed record RedeemRequestModel(string Code);
