+++ src/web/Tomeshelf.Web/Models/Shift/RedeemResponseModel.cs
@@ -1,0 +1,5 @@
+using System.Collections.Generic;
+
+namespace Tomeshelf.Web.Models.Shift;
+
+public sealed record RedeemResponseModel(RedeemSummaryModel Summary, IReadOnlyList<RedeemResultModel> Results);
