+++ src/web/Tomeshelf.Web/Models/Shift/RedeemResultModel.cs
@@ -1,0 +1,3 @@
+namespace Tomeshelf.Web.Models.Shift;
+
+public sealed record RedeemResultModel(int AccountId, string Email, string Service, bool Success, RedeemErrorCode? ErrorCode, string? Message);
