+++ src/web/Tomeshelf.Web/Models/Shift/RedeemSummaryModel.cs
@@ -1,0 +1,3 @@
+namespace Tomeshelf.Web.Models.Shift;
+
+public sealed record RedeemSummaryModel(int Total, int Succeeded, int Failed);
