+++ src/web/Tomeshelf.Web/Services/FitbitApi.cs
@@ -150,60 +150,3 @@ public sealed class FitbitApi : IFitbitApi
         return (statusCode == HttpStatusCode.Redirect) || (statusCode == HttpStatusCode.RedirectKeepVerb) || (statusCode == HttpStatusCode.RedirectMethod) || (statusCode == HttpStatusCode.MovedPermanently) || (statusCode == HttpStatusCode.SeeOther) || (statusCode == HttpStatusCode.TemporaryRedirect) || (statusCode == HttpStatusCode.PermanentRedirect);
     }
 }
-
-public sealed class FitbitAuthorizationRequiredException : Exception
-{
-    public FitbitAuthorizationRequiredException(Uri location) : base("Fitbit authorization is required.")
-    {
-        Location = location;
-    }
-
-    public Uri Location { get; }
-}
-
-public sealed class FitbitBackendUnavailableException : Exception
-{
-    public FitbitBackendUnavailableException(string message) : base(BuildMessage(message)) { }
-
-    private static string BuildMessage(string raw)
-    {
-        if (string.IsNullOrWhiteSpace(raw))
-        {
-            return "Fitbit service is unavailable. Please try again in a moment.";
-        }
-
-        var trimmed = raw.Trim();
-        if (trimmed.StartsWith("<", StringComparison.Ordinal))
-        {
-            return "Fitbit service is unavailable. Please try again in a moment.";
-        }
-
-        if (trimmed.StartsWith("{", StringComparison.Ordinal))
-        {
-            try
-            {
-                using var document = JsonDocument.Parse(trimmed);
-                if ((document.RootElement.ValueKind == JsonValueKind.Object) && document.RootElement.TryGetProperty("message", out var messageElement) && (messageElement.ValueKind == JsonValueKind.String))
-                {
-                    var parsedMessage = messageElement.GetString();
-                    if (!string.IsNullOrWhiteSpace(parsedMessage))
-                    {
-                        return parsedMessage.Trim();
-                    }
-                }
-            }
-            catch (JsonException) { }
-        }
-
-        if (trimmed.StartsWith("\"", StringComparison.Ordinal) && trimmed.EndsWith("\"", StringComparison.Ordinal))
-        {
-            try
-            {
-                trimmed = JsonSerializer.Deserialize<string>(trimmed) ?? trimmed;
-            }
-            catch (JsonException) { }
-        }
-
-        return trimmed;
-    }
-}
