+++ tests/services/mcm/Tomeshelf.MCM.Infrastructure.Tests/TestUtilities/LogEntry.cs
@@ -1,0 +1,16 @@
+using Microsoft.Extensions.Logging;
+
+namespace Tomeshelf.MCM.Infrastructure.Tests.TestUtilities;
+
+public class LogEntry
+{
+    public LogLevel LogLevel { get; set; }
+
+    public EventId EventId { get; set; }
+
+    public object? State { get; set; }
+
+    public Exception? Exception { get; set; }
+
+    public string? Message { get; set; }
+}
